pyproject.toml
[project]
name = "research-assistant-sys"
version = "0.1.0"
description = "Research Assistant Sys: Agno-based orchestration with pluggable experts (paper structure extraction, etc.)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno",
    "python-dotenv",
    "openai",
    "arxiv",
    "pypdf",
    "numpy",
    "scipy",
    "pydantic>=2.0",
    "fastapi>=0.135.3",
]

[project.optional-dependencies]
marker = ["marker-pdf"]
nougat = ["nougat-ocr"]
remote-ocr = ["pymupdf"]
anthropic = ["anthropic"]
google = ["google-genai"]
all = [
    "marker-pdf",
    "nougat-ocr",
    "pymupdf",
    "anthropic",
    "google-genai",
]

[project.scripts]
researchsys = "main:main"
optibench = "main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["cli*", "experts*", "foundation*", "interface*", "orchestrator*"]
